Moved unites test into its own directory.
