+++ src/applications/my-education-benefits/tests/units/components/GetFormHelp.unit.spec.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { expect } from 'chai';
-// import { isReactComponent } from 'platform/utilities/ui';
-import GetFormHelp from '../components/GetFormHelp';
+import GetFormHelp from '../../../components/GetFormHelp';
 
 describe('GetFormHelp component', () => {
   it('renders the Get Help footer', () => {
